requirements.txt
albumentations==1.3.1
dash==2.15.0
dash_bootstrap_components==1.5.0
numba==0.59.0
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.1
plotly==5.18.0
python-dotenv==1.0.1
scipy==1.12.0
segmentation_models_pytorch==0.3.3
skimage==0.0
tifffile==2024.1.30
torch==2.2.0+cu121
tqdm==4.66.2
